refactor(A3): split global.ts god-module; break global.ts ⇄ wos.ts cycle
Extract 3 focused modules from the 1047-line store/global.ts:
  - app-api.ts: getApi() host bridge accessor (13 LOC)
  - config-signals.ts: fullConfigAtom / settingsAtom / hasCustomAIPresetsAtom (23 LOC)
  - block-atom-cache.ts: per-block/tab SolidJS signal caches and settings memos (260 LOC)

All symbols re-exported from global.ts — no call-sites change import paths.

Cycle fix: wos.ts imported getApi from global.ts while global.ts imported
WOS from wos.ts. wos.ts now imports from app-api.ts instead.

Leaf violations fixed:
- util/logger.ts: getApi from @/app/store/app-api (not the god-module)
- layout/lib/layoutModel.ts: getSettingsKeyAtom from @/app/store/block-atom-cache;
  unused `atoms` import removed

global.ts: 1047 → 821 LOC (-226). A1 contract test: 4/4 pass. Zero new TS errors.
Part of the A1-A15 architecture refactor board (A3).
